Use chef to generate Vagrant file, chef recipes.
Infrastructure's goal is to keep the Way development environments consistent across the repos of the organization.  It aims to do so by extracting their provisioning to a single place.  Hand in hand with this goal is automating the process of creating a new repo in the organization.

There is a natural trade-off between similarity (for portable UX) and diversity (for achieving required function) of org repos.  Infrastructure takes the following stance:
* All repos will avoid a thick host OS tool chain by sand-boxing development within vagrant.
* All repos will use ruby and its ecosystem (e.g. rake) for tooling where possible.
* All repos will automate development environment provisioning with chef.
* Infrastructure will clearly divide maintenance responsibility of common files, and all repos will respect this division of responsibility.

The first point seeks to limit host OS dependencies to:
* git
* vagrant
* text editor of choice

The goal of the last point is to avoid code rot introduced by projects being created at various times in the face of evolving project templates.  When a significantly complex file changes in a way that all repos should consume, infrastructure will roll the change out (e.g. as an updated chef template.)

Infrastructure makes an effort to allow client repositories to retain ownership of most files.  Infrastructure implements two strategies for managing the portions of those files that it owns.  First, it tries to offer an inclusion mechanism of some sort either native to the file format or via a chef partial template.  When inclusion is not appropriate, it falls back to validating client-owned files at provision time to ensure that required features are present (see `project_spec.rb`.)

Infrastructure explicitly does not require:
* Committing to a host OS (in theory)
* Committing to an editor, though [editorconfig][editorconfig] users will benefit.
* Committing to or automating the configuration of external services such as travis CI.

[editorconfig]: http://editorconfig.org/

The `install` script (meant for `curl | sh`) fetches the latest [github release][github-releases] of infrastructure, which generates a Vagrantfile and drops into chef provisioning to create the rest of the project template.

[github-releases]: https://help.github.com/articles/about-releases/