Drop subsection `(c)` of section 4 (`NOTICE` file preservation).
The `NOTICE` file serves to contain legally-required notices that do not fit anywhere else.  This license isn't required to include such a clause for those requirements to have power (nor can it negate it).  Instead it simply tries to enshrine a particular file to add a layer of.. redundancy?

In other words it is over-engineered.