GPL3.
Put the text itself in `GPL-3`.  `LICENSE` is now a one-liner declaring the GPL3, with the "or any later version" option.

`src/check-boilerplate` (softlinked in `bin`) is a tool to check the health of boilerplate in the project.  Currently it looks for copyright declarations, current end years of those declarations, and a line from the GPL boilerplate.  I imagine later it will also grow mode lines and encoding hints.  It is run as part of `layout.bats`.

closes #19.