Terms
Here is a term datatype for a basic type theory.  There is no typing relation yet (we
don't immediately reach for a "pure type theory"), nor equality nor sigma types.  We just
have introduction/elimination for abstractions and products, and stratified universe
constants.

Variables are handled in the "locally nameless" style, where free variables are identified
as `atom`s and bound variables are de Bruijn indices.  To ensure that bound variables
refer to actual binders, the datatype is actually `preterm` instead of `term`, which is
reserved for `preterms` that satisfy a well-formedness constraint named `locally_closed`.

Closes #23